feat(controller): move configuration package to pkg/ for public consumption
  Move internal/configuration to pkg/configuration to make the
  configuration loading API available to external consumers. Add package
  documentation with an API stability disclaimer. Update all internal
  import paths accordingly.
